LayoutTests/js/globalThis-is-window-expected.txt

This test checks that the globalThis property is the same as the window object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS globalThis is window
PASS successfullyParsed is true

TEST COMPLETE
PASS globalThis is window
